+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+* Switch from using `format!` to `String.push` and `String.push_str`. +/- 8% performance improvement.
 
 ## [0.2.1] 2020-04-08
 ### Fixed